feat(monitoring): Scout aktiviert + INT4-Decode-Fix — Breitensammlung DE-Deadlock-Streamer
Der Scout (entdeckt live DE-Deadlock-Streamer inkl. Nicht-Partner, nimmt sie ins
Monitoring → Poll sammelt live/viewer/session/peak ohne Grant) war seit dem Cutover
per TB_SCOUT_ENABLED gegated AUS → Roster auf ~56 Partner geschrumpft, keine neuen
Nicht-Partner seit 16.06. Flag default auf 1 gesetzt (Python-Paritaet). Beim Aktivieren
INT4-Decode-Bug gefunden+gefixt: Existenz-Check 'SELECT 1' (INT4) wurde als i64 dekodiert
→ Upsert scheiterte fuer jeden entdeckten Streamer. Jetzt i32. Live verifiziert: Scout-
Zyklus laeuft sauber (0 mismatched-types), Discovery aktiv.

+++ rust/crates/tb-monitoring/src/scout.rs
@@ -130,7 +130,7 @@ impl ScoutRepository {
     /// Trägt einen neuen Monitoring-only-Streamer ein. Gibt `true` zurück wenn
     /// er tatsächlich neu war (nicht nur ein Konflikt-Update).
     pub async fn upsert_monitored(&self, login: &str, user_id: &str) -> Result<bool, sqlx::Error> {
-        let existing: Option<i64> = sqlx::query_scalar(
+        let existing: Option<i32> = sqlx::query_scalar(
             "SELECT 1 FROM twitch_streamers WHERE LOWER(twitch_login) = LOWER($1) LIMIT 1",
         )
         .bind(login)
